scripts/xtensa-build-zephyr.py: Correct XTENSA_SYSTEM path
The current build fails if building with XTENSA_TOOLS_ROOT set to a
non-default Cadence tool directly.  The proximate failure is that
"xt-clang --version" produces an error at zephyr configure time (which
is hidden by default, sigh) that it can't find the core in the
specified XTENSA_SYSTEM registry.

When you run "xt-clang" without arguments and without XTENSA_CORE set,
it will helpfully point to the list of installed cores, but they're at
a different spot than this script is using.  Basically the core
registry is installed with the "tools", not the "builds", and isn't
specific to a given core.  Whch makes sense, I guess.

The directories don't otherwise seem compatible, I don't think this
ever worked?  Maybe I'm the only one using a XTENSA_TOOLS_ROOT
setting?

Signed-off-by: Andy Ross <andyross@google.com>