demux: Ignore inactive output buffers
It may happen that a demux is configured to write to a buffer in a
stopped pipeline (consider echo cancellation when the mic is not in
use: the output reference stream has nowhere to go).  Strictly this
happens because the pipeline ID for the output buffer is unset, as it
wasn't traversed during its own pipe's prepare step.

Just drop these streams dynamically at process() time, as the target
pipeline may be started up at arbitrary moments.  When it is, we'll
begin emitting output at the first process callback as desired.

Signed-off-by: Andy Ross <andyross@google.com>