pipeline: Fix reset propagation logic
Behaviorally, this is a partial revert of commit 2f5bf3cc92cb
("pipeline: change pipe reset propagation condition").  That patch
added logic to propagate resets[1] to components on other pipelines,
as long as they were on the same scheduler component (i.e. would be
updated synchronously).

But it got the precedence wrong: the pre-existing code would REFUSE to
propagate to components with an opposite direction (i.e. capture to
playback), but now it would do so as the components shared a
scheduler.  This breaks cases like AEC where playback may be shut down
while capture is running; we don't want to reset the AEC component on
the capture pipeline just because the playback demux that feeds it got
reset!

[1] When using IPC3. In the current tree, IPC4 never propagates across
    pipelines, making this issue moot.

Signed-off-by: Andy Ross <andyross@google.com>