google_rtc_audio_processing: Handle inactive reference input
Close the loop on "optional pipeline" state management.  When the
component connected to the reference input is inactive, synthesize
zeros in the buffer being fed to echo cancellation processing.

Now the capture pipe is no longer necessary and no host DMA needs to
be happening on the playback channel just to feed AEC.  You can start
and stop playback at any time and the stream will accommodate.

Note that mic startup happens on the host's schedule and isn't
synchronous, so there's a case when exercising this code where the
reference stream can be a few 10ms blocks ahead of the mic when it
finally arrives.  Add a "catch up" check to discard stale samples.

Signed-off-by: Andy Ross <andyross@google.com>