google_rtc_audio_processing: Support multiple sample formats
The processing code is now almost entirely sample-format-independent,
so finish the job:

Add kconfigs to control the sample formats of both the reference and
capture streams.  S32_LE and S16_LE are currently supported.

Use the float32 variant of the AEC library instead of the int16
wrapper.  The internal API has always been floating point, this just
moves the conversion into the wrapper code.  Longer term this is
probably a good thing for multiprocessor performance as it allows the
conversions to be decoupled from the analysis code and moved to
another component on another core.  (Not in this version though, it's
still synchronous here).

Signed-off-by: Andy Ross <andyross@google.com>