google_rtc_audio_processing: Don't mess with source/sink formats
This code was specific to IPC4, and used a IPC4-specific API to try to
set formats on the connected sources/sinks.  Which seems questionable?
The format of external data isn't ours to define, the topology sets it
and we should be set up compatibly.

It was also disabled in the mtl-007-topic-branch with a comment of
"does not work", so let's just remove it until it's clear why it's
needed.

Signed-off-by: Andy Ross <andyross@google.com>