google_rtc_audio_processing: Work around IPC4 stream setup issues
IPC4 fails to set up stream metadata on the buffer objects at pipeline
creation time (see Issue #8639).  Traditionally components have been
responsible for doing this at prepare time via calling
ipc4_update_source/sink_format() to set values on the (incompletely
initialized) streams.

In particular, on MTL the reference input stream is presented with an
incorrect channel count (4, matching the capture stream; but the
reference/playback pipeline has only two channels) and needs to be
corrected.

Also note: The "pin" index (the value of the pin_index field of the
ipc4_in/output_pin_format structs seen in module config at init()
time) and the "source" index (the ordering of the sources[] argument
to prepare()/process()) ARE PRESENTED BACKWARDS!  And I don't see any
API to tell me which is which (note that the ordering of the connected
buffers in the comp_dev sink/source lists is a THIRD convention, and
matches sources[]/sinks[]).

Signed-off-by: Andy Ross <andyross@google.com>

[DNM] Google AEC: minimize workaround

This is down to just one line now with existing (still in review) PRs
applied