module_adapter: Remove BASE_CONFIG_EXT attribute
There was only ever one consumer for this attribute, and it didn't
actually work.  Just remove it, the ipc4 connection code is now using
values directly out of module config.

The BASE_CONFIG attribute does have a few components that are
implementing their own version, so it can't be immediately removed.

Signed-off-by: Andy Ross <andyross@google.com>