drivers/mtk/afe: Update timestamp API
This DMA driver was using timer_get_system() to generate timestamps,
which is the Xtensa CCOUNT timer in legacy SOF but doesn't exist at
all in Zephyr (well, you can read the SR directly).  Use
sof_cycle_get_64(), which is the portable API based on the clock
hardware.  Other DMA drivers do just fine with this.

Signed-off-by: Andy Ross <andyross@google.com>