ipc wired up (but... boom?)
Send the FW_READY IPC back.  And the whole device reboots!  But not if
I do either operation in isolation: writing the mailbox works fine
(kernel sees nothing and times out).  Sending the IPC causes a kernel
parse error.  So somehow the contents of the FW_READY command are
panicking the kernel, fun!