src/ipc3: Don't fail on trace messages when CONFIG_TRACE=n
Zephyr builds don't use the older trace protocol, but for now still
need to work with IPC3 kernel layers that expect it to be present.
Just return a noop success from trace messages when not configured,
otherwise the -EINVAL will propagate to the other side and abort the
firmware load.

Signed-off-by: Andy Ross <andyross@google.com>