src/include/component_ext: Declare Zephyr log module
This header can be included in Zephyr builds in a context where the no
log module has been declared.  (Just in general, zephyr logging
requires a local context and doesn't work well in inline functions
that live in headers; macros work better.) Declare it locally to avoid
a build failure.

Signed-off-by: Andy Ross <andyross@google.com>