include/interrupt.h: Zephyr build fixups
This header may be included in situations where there is no log
context declared, but still needs to expand a tr_err() macro.  Declare
the "zephyr" context as a default.  (Though really this is not a
runtime error and should probably be handled with a BUILD_ASSERT()).

Also, there was a growing "enumerate all the Zephyr platforms"
preprocessor #if expression that can be simplified to
"defined(__ZEPHYR__)"

Signed-off-by: Andy Ross <andyross@google.com>