zephyr/app: Clean up Zephyr kconfig defaults
Things had gotten a little messy here.  The CPU_MASK_PIN_ONLY and
SMP_BOOT_DELAY options are in fact not generic to all Zephyr builds;
they both depend on SMP, and in fact don't exist otherwise, producing
a build failure if the board defconfig doesn't turn it off.  Create an
application "Kconfig" file (which SOF hasn't used so far), which
allows the dependencies for these settings to be more cleanly
expressed.

Similarly CONFIG_TRACE=y won't build with Zephyr, and all boards were
turning it off independently.  Move that to the prj.conf level.

And likewise CONFIG_ZEPHYR_LOG, while not strictly required for a
Zephyr build, is pervasively enabled (except on ACP6?  What does that
device use for logging?), so it belongs in the global file.

Signed-off-by: Andy Ross <andyross@google.com>