math: Fix dependencies in filter HiFi selection
The intent of the dependencies on this choice seems to be that they
apply only to certain components so the selection should be hidden if
they aren't in use.

But the logic was wrong: instead of enabling it if any component was
enabled that needed it, it would hide the selection unless ALL such
components were enabled.  Which doesn't build.  Fix.

(Also I note the comment implies that other kconfigs should probably
be in the list too, but not adding them out of conservatism as this
is, again, just a hygiene thing.)

Signed-off-by: Andy Ross <andyross@google.com>